requirements.txt
Django==2.2.5
PyMySQL==0.9.3
djangorestframework==3.10.3
